Several changes to BIDS file structure
 - Split bids/types.js into several files.
 - Move curly brace mapping generation to BidsSidecar.
 - Fix test issue.
 - Don't include columns spliced through curly braces in combined string.
