--- bids/types.js
@@ -1,422 +1,0 @@
-import { sidecarValueHasHed } from './utils'
-import { generateIssue, Issue, IssueError } from '../common/issues/issues'
-import { parseTSV, convertParsedTSVData } from './tsvParser'
-import { parseHedString } from '../parser/main'
-
-/**
- * Base class for BIDS data.
- * @deprecated Will be removed in v4.0.0.
- */
-class BidsData {
-  /**
-   * A mapping from unparsed HED strings to ParsedHedString objects.
-   * @deprecated Will be removed in v4.0.0.
-   * @type {Map<string, ParsedHedString>}
-   */
-  parsedStringMapping
-  /**
-   * A Mapping from definition names to their associated ParsedHedGroup objects.
-   * @deprecated Will be removed in v4.0.0.
-   * @type {Map<string, ParsedHedGroup>}
-   */
-  definitions
-  /**
-   * A list of HED validation issues.
-   * This will be converted to BidsIssue objects later on.
-   * @deprecated Will be removed in v4.0.0.
-   * @type {Issue[]}
-   */
-  hedIssues
-
-  constructor() {
-    this.parsedStringMapping = new Map()
-    this.definitions = new Map()
-    this.hedIssues = []
-  }
-}
-
-/**
- * A BIDS file.
- */
-class BidsFile extends BidsData {
-  /**
-   * The name of this file.
-   * @type {string}
-   */
-  name
-  /**
-   * The file object representing this file data.
-   * This is used to generate BidsIssue objects.
-   * @type {object}
-   */
-  file
-
-  constructor(name, file) {
-    super()
-    this.name = name
-    this.file = file
-  }
-}
-
-/**
- * A BIDS JSON file.
- */
-export class BidsJsonFile extends BidsFile {
-  /**
-   * This file's JSON data.
-   * @type {object}
-   */
-  jsonData
-
-  constructor(name, jsonData, file) {
-    super(name, file)
-    this.jsonData = jsonData
-  }
-}
-
-/**
- * A BIDS TSV file.
- */
-export class BidsTsvFile extends BidsFile {
-  /**
-   * This file's parsed TSV data.
-   * @type {Map<string, string[]>}
-   */
-  parsedTsv
-  /**
-   * HED strings in the "HED" column of the TSV data.
-   * @type {string[]}
-   */
-  hedColumnHedStrings
-  /**
-   * The list of potential JSON sidecars.
-   * @type {string[]}
-   */
-  potentialSidecars
-  /**
-   * The pseudo-sidecar object representing the merged sidecar data.
-   * @type {BidsSidecar}
-   */
-  mergedSidecar
-  /**
-   * The extracted HED data for the merged pseudo-sidecar.
-   * @type {Map<string, string|Object<string, string>>}
-   */
-  sidecarHedData
-
-  /**
-   * Constructor.
-   *
-   * @todo This interface is provisional and subject to modification in version 4.0.0.
-   *
-   * @param {string} name The name of the TSV file.
-   * @param {{headers: string[], rows: string[][]}|Map<string, string[]>|string} tsvData This file's TSV data.
-   * @param {object} file The file object representing this file.
-   * @param {string[]} potentialSidecars The list of potential JSON sidecars.
-   * @param {object} mergedDictionary The merged sidecar data.
-   */
-  constructor(name, tsvData, file, potentialSidecars = [], mergedDictionary = {}) {
-    super(name, file)
-    let parsedTsvData
-    if (typeof tsvData === 'string') {
-      parsedTsvData = parseTSV(tsvData)
-    } else if (tsvData === Object(tsvData)) {
-      parsedTsvData = convertParsedTSVData(tsvData)
-    } else {
-      parsedTsvData = tsvData
-    }
-    this.parsedTsv = parsedTsvData
-    this.potentialSidecars = potentialSidecars
-
-    this.mergedSidecar = new BidsSidecar(name, mergedDictionary, null)
-    this.sidecarHedData = this.mergedSidecar.hedData
-    this._parseHedColumn()
-  }
-
-  _parseHedColumn() {
-    const hedColumn = this.parsedTsv.get('HED')
-    if (hedColumn === undefined) {
-      this.hedColumnHedStrings = []
-    } else {
-      this.hedColumnHedStrings = hedColumn.map((hedCell) => (hedCell && hedCell !== 'n/a' ? hedCell : ''))
-    }
-  }
-}
-
-/**
- * A BIDS events.tsv file.
- */
-export class BidsEventFile extends BidsTsvFile {
-  /**
-   * Constructor.
-   *
-   * @todo This interface is subject to modification in version 4.0.0.
-   *
-   * @param {string} name The name of the event TSV file.
-   * @param {string[]} potentialSidecars The list of potential JSON sidecars.
-   * @param {object} mergedDictionary The merged sidecar data.
-   * @param {{headers: string[], rows: string[][]}|string} tsvData This file's TSV data.
-   * @param {object} file The file object representing this file.
-   */
-  constructor(name, potentialSidecars, mergedDictionary, tsvData, file) {
-    super(name, tsvData, file, potentialSidecars, mergedDictionary)
-  }
-}
-
-/**
- * A BIDS TSV file other than an events.tsv file.
- */
-export class BidsTabularFile extends BidsTsvFile {
-  /**
-   * Constructor.
-   *
-   * @todo This interface is subject to modification in version 4.0.0.
-   *
-   * @param {string} name The name of the TSV file.
-   * @param {string[]} potentialSidecars The list of potential JSON sidecars.
-   * @param {object} mergedDictionary The merged sidecar data.
-   * @param {{headers: string[], rows: string[][]}|string} tsvData This file's TSV data.
-   * @param {object} file The file object representing this file.
-   */
-  constructor(name, potentialSidecars, mergedDictionary, tsvData, file) {
-    super(name, tsvData, file, potentialSidecars, mergedDictionary)
-  }
-}
-
-export class BidsSidecar extends BidsJsonFile {
-  /**
-   * The extracted HED data for this sidecar.
-   * @type {Map<string, string|Object<string, string>>}
-   */
-  hedData
-  /**
-   * The parsed HED data for this sidecar.
-   * @type {Map<string, ParsedHedString|Map<string, ParsedHedString>>}
-   */
-  parsedHedData
-  /**
-   * The extracted HED value strings.
-   * @type {string[]}
-   */
-  hedValueStrings
-  /**
-   * The extracted HED categorical strings.
-   * @type {string[]}
-   */
-  hedCategoricalStrings
-
-  /**
-   * Constructor.
-   *
-   * @param {string} name The name of the sidecar file.
-   * @param {Object} sidecarData The raw JSON data.
-   * @param {Object} file The file object representing this file.
-   */
-  constructor(name, sidecarData = {}, file) {
-    super(name, sidecarData, file)
-
-    this._filterHedStrings()
-    this._categorizeHedStrings()
-  }
-
-  _filterHedStrings() {
-    const sidecarHedTags = Object.entries(this.jsonData)
-      .map(([sidecarKey, sidecarValue]) => {
-        if (sidecarValueHasHed(sidecarValue)) {
-          return [sidecarKey, sidecarValue.HED]
-        } else {
-          return []
-        }
-      })
-      .filter((x) => x.length > 0)
-    this.hedData = new Map(sidecarHedTags)
-  }
-
-  _categorizeHedStrings() {
-    this.hedValueStrings = []
-    this.hedCategoricalStrings = []
-    for (const sidecarValue of this.hedData.values()) {
-      if (typeof sidecarValue === 'string') {
-        this.hedValueStrings.push(sidecarValue)
-      } else {
-        this.hedCategoricalStrings.push(...Object.values(sidecarValue))
-      }
-    }
-  }
-
-  /**
-   * Parse this sidecar's HED strings within the sidecar structure.
-   *
-   * The parsed strings are placed into {@link parsedHedData}.
-   *
-   * @param {Schemas} hedSchemas The HED schema collection.
-   * @returns {Issue[]} Any issues found.
-   */
-  parseHedStrings(hedSchemas) {
-    this.parsedHedData = new Map()
-    const issues = []
-    for (const [key, value] of this.hedData) {
-      issues.push(...this._parseSidecarKey(key, value, hedSchemas))
-    }
-    return issues
-  }
-
-  _parseSidecarKey(key, data, hedSchemas) {
-    if (typeof data === 'string') {
-      return this._parseHedString(this.parsedHedData, key, data, hedSchemas)
-    } else if (data !== Object(data)) {
-      return [generateIssue('illegalSidecarHedType', { key: key, file: this.name })]
-    }
-    const issues = []
-    const keyMap = new Map()
-    for (const [value, string] of Object.entries(data)) {
-      issues.push(...this._parseHedString(keyMap, value, string, hedSchemas))
-    }
-    this.parsedHedData.set(key, keyMap)
-    return issues
-  }
-
-  _parseHedString(map, key, string, hedSchemas) {
-    const [parsedString, parsingIssues] = parseHedString(string, hedSchemas)
-    map.set(key, parsedString)
-    return Object.values(parsingIssues).flat()
-  }
-
-  /**
-   * The extracted HED strings.
-   * @returns {string[]}
-   */
-  get hedStrings() {
-    return this.hedValueStrings.concat(this.hedCategoricalStrings)
-  }
-
-  /**
-   * An alias for {@link jsonData}.
-   * @returns {Object}
-   */
-  get sidecarData() {
-    return this.jsonData
-  }
-}
-
-/**
- * Fallback default dataset_description.json file.
- * @deprecated Will be removed in v4.0.0.
- * @type {BidsJsonFile}
- */
-const fallbackDatasetDescription = new BidsJsonFile('./dataset_description.json', null)
-
-export class BidsDataset extends BidsData {
-  /**
-   * The dataset's event file data.
-   * @type {BidsEventFile[]}
-   */
-  eventData
-  /**
-   * The dataset's sidecar data.
-   * @type {BidsSidecar[]}
-   */
-  sidecarData
-  /**
-   * The dataset's dataset_description.json file.
-   * @type {BidsJsonFile}
-   */
-  datasetDescription
-  /**
-   * The dataset's root directory as an absolute path.
-   * @type {string|null}
-   */
-  datasetRootDirectory
-
-  constructor(eventData, sidecarData, datasetDescription = fallbackDatasetDescription, datasetRootDirectory = null) {
-    super()
-    this.eventData = eventData
-    this.sidecarData = sidecarData
-    this.datasetDescription = datasetDescription
-    this.datasetRootDirectory = datasetRootDirectory
-  }
-}
-
-const bidsHedErrorCodes = new Set([104, 106, 107])
-
-export class BidsIssue {
-  /**
-   * The BIDS issue code.
-   * @type {number}
-   */
-  code
-  /**
-   * The file associated with this issue.
-   * @type {Object}
-   */
-  file
-  /**
-   * The evidence for this issue.
-   * @type {string}
-   */
-  evidence
-
-  constructor(issueCode, file, evidence) {
-    this.code = issueCode
-    this.file = file
-    this.evidence = evidence
-  }
-
-  /**
-   * Whether this issue is an error.
-   * @returns {boolean}
-   */
-  isError() {
-    return bidsHedErrorCodes.has(this.code)
-  }
-
-  static generateInternalErrorPromise(error) {
-    return Promise.resolve([new BidsIssue(106, null, error.message)])
-  }
-}
-
-export class BidsHedIssue extends BidsIssue {
-  /**
-   * The HED Issue object corresponding to this object.
-   * @type {Issue}
-   */
-  hedIssue
-
-  constructor(hedIssue, file) {
-    super(hedIssue.level === 'warning' ? 105 : 104, file, hedIssue.message)
-
-    this.hedIssue = hedIssue
-  }
-
-  /**
-   * Convert one or more HED issues into BIDS-compatible issues.
-   *
-   * @param {IssueError|Issue[]} hedIssues One or more HED-format issues.
-   * @param {Object} file A BIDS-format file object used to generate {@link BidsHedIssue} objects.
-   * @param {Object?} extraParameters Any extra parameters to inject into the {@link Issue} objects.
-   * @returns {BidsHedIssue[]} The passed issue(s) in BIDS-compatible format.
-   */
-  static fromHedIssues(hedIssues, file, extraParameters) {
-    if (hedIssues instanceof IssueError) {
-      return [BidsHedIssue.fromHedIssue(hedIssues.issue, file, extraParameters)]
-    } else {
-      return hedIssues.map((hedIssue) => BidsHedIssue.fromHedIssue(hedIssue, file, extraParameters))
-    }
-  }
-
-  /**
-   * Convert a single HED issue into a BIDS-compatible issue.
-   *
-   * @param {Issue} hedIssue One HED-format issue.
-   * @param {Object} file A BIDS-format file object used to generate a {@link BidsHedIssue} object.
-   * @param {Object?} extraParameters Any extra parameters to inject into the {@link Issue} object.
-   * @returns {BidsHedIssue} The passed issue in BIDS-compatible format.
-   */
-  static fromHedIssue(hedIssue, file, extraParameters) {
-    if (extraParameters) {
-      Object.assign(hedIssue.parameters, extraParameters)
-      hedIssue.generateMessage()
-    }
-    return new BidsHedIssue(hedIssue, file)
-  }
-}
